coverity#706062 Unintended sign extension

+++ starmath/source/node.cxx
@@ -2544,7 +2544,7 @@ void SmMatrixNode::Arrange(const OutputDevice &rDev, const SmFormat &rFormat)
     }
 
     // norm distance from which the following two are calcutated
-    const int  nNormDist = 3 * GetFont().GetSize().Height();
+    const long  nNormDist = 3 * GetFont().GetSize().Height();
 
     // define horizontal and vertical minimal distances that separate
     // the elements
